[sysfs] Make the timespec32 narrowing explicit (-Wshorten-64-to-32)
modules/sysfs/sysfs.c:127 assigned a time_t value (long long on
musl) into k_timespec32::tv_sec (s32). The narrowing is
deliberate -- struct k_stat64 documents its Y2038 limitation in
sys_types.h -- but the implicit conversion trips
-Wshorten-64-to-32 under clang_wconv's -Werror. Add an explicit
(s32) cast.

Pre-existing issue (line dates back to 2020); only surfaced now
because earlier inttypes.h-related failures prevented
clang_wconv from reaching this TU.

+++ modules/sysfs/sysfs.c
@@ -124,7 +124,14 @@ sysfs_stat(struct mnt_fs *fs, vfs_inode_ptr_t i, struct k_stat64 *statbuf)
    statbuf->st_gid = 0; /* root */
    statbuf->st_blksize = PAGE_SIZE;
    statbuf->st_blocks = 0;
-   statbuf->st_ctim.tv_sec = d->wrt_time;
+   /*
+    * Narrowing to k_timespec32::tv_sec is part of the existing
+    * Y2038 limitation of struct k_stat64 — see the comment on
+    * the struct declaration in include/tilck/kernel/sys_types.h.
+    * Make the cast explicit so -Wshorten-64-to-32 (clang_wconv)
+    * doesn't trip on what's a deliberate truncation.
+    */
+   statbuf->st_ctim.tv_sec = (s32)d->wrt_time;
    statbuf->st_mtim = statbuf->st_ctim;
    statbuf->st_atim = statbuf->st_mtim;
